requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0

# Vector database and embeddings
chromadb>=0.5.0
openai>=1.50.0

# Text processing
tiktoken>=0.8.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data processing and similarity
numpy>=1.24.0
scikit-learn>=1.3.0

# Database and encryption for user management
sqlalchemy>=2.0.23
cryptography>=41.0.7

# Web framework
streamlit>=1.39.0
streamlit-oauth>=0.1.5

# HTTP requests
requests>=2.31.0

# Confluence integration
atlassian-python-api>=3.41.0

# Environment management
python-dotenv>=1.0.0
pytz>=2023.3

# Database (PostgreSQL)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
